HORNETQ-1420 Log Bridge connect retry WARN once
When cluster connection bridge fails to connect to the cluster node it
will log retry messages at WARN level until it does so.  Polluting the
logs with lots of WARN messages.  This patch changes the logging so that
retry message is logged at level WARN only once and adds in a separate
DEBUG message for sunsequent reconnect attempts.